Rename `run-medley` args, pass unknowns through.
This allows for a more consistent command-line interface that doesn't
have to duplicate lde's argument parsing. We handle four arguments that
are inconvenient when calling lde directly or that require setting
environment variables, and then the rest is passed through.

At this point, `run-medley-mac` shouldn't be necessary, but I haven't
tested this revised script on a Mac yet, so I'm leaving it be for now.